// src/app/shared/models/matrizEvidencia.ts
export class MatrizEvidencia {
    data?: any;
    idEvidencia?: number;
    idMatrizRiesgo?: number;
    idDetaMatrizRiesgo?: number;
    nombreOriginalArchivo?: string;
    nombreArchivo?: string;
    pesoArchivo?: string;
    rutaArchivo?: string;
    tipoEvidencia?: number;
    usuarioCreacion?: string;
    ipCreacion?: string;
}